+++ src/main/kotlin/com/doubleu/kotlintrader/model/Ort.kt
@@ -3,7 +3,7 @@ package com.doubleu.kotlintrader.model
 import com.doubleu.kotlintrader.database.Entity
 import com.doubleu.kotlintrader.extensions.valueOf
 
-class Ort(override val id: Int) : Entity() {
+class Ort(override val id: Long) : Entity() {
 
     var name: String? by delegate(this::name)
     val nameProperty = property(this::name)
