+++ src/main/kotlin/com/doubleu/kotlintrader/model/Schiff.kt
@@ -2,15 +2,15 @@ package com.doubleu.kotlintrader.model
 
 import com.doubleu.kotlintrader.database.Entity
 
-class Schiff(override val id: Int) : Entity() {
+class Schiff(override val id: Long) : Entity() {
 
-    var ort_id: Int? by delegate(this::ort_id)
+    var ort_id: Long? by delegate(this::ort_id)
     val ort_idProperty = property(this::ort_id)
 
     var ort: Ort? by mutableReference(this::ort, this::ort_id)
     val ortProperty = property(this::ort)
 
-    var trader_id: Int? by delegate(this::trader_id)
+    var trader_id: Long? by delegate(this::trader_id)
     val trader_idProperty = property(this::trader_id)
 
     var trader: Trader? by mutableReference(this::trader, this::trader_id)
