+++ src/main/kotlin/com/doubleu/kotlintrader/model/Ware.kt
@@ -2,7 +2,7 @@ package com.doubleu.kotlintrader.model
 
 import com.doubleu.kotlintrader.database.Entity
 
-class Ware(override val id: Int) : Entity() {
+class Ware(override val id: Long) : Entity() {
 
     var name: String? by delegate(this::name)
     val nameProperty = property(this::name)
